Fixes George comments
